requirements.txt
Django>=4.0.3,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2>=2.9.3,<2.10.0
Pillow>=9.1.0,<9.2.0

flake8>=4.0.1,<4.1.0